requirements.txt
amqp==5.3.1
asgiref==3.8.1
async-timeout==5.0.1
backports.zoneinfo==0.2.1
billiard==4.2.1
celery==5.5.2
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
django==4.2.21
django-cors-headers==4.4.0
django-decouple==2.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.10
inflection==0.5.1
kombu==5.5.3
packaging==25.0
prompt-toolkit==3.0.51
PyJWT==2.9.0
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
redis==6.1.0
six==1.17.0
sqlparse==0.5.3
typing-extensions==4.13.2
uritemplate==4.1.1
vine==5.1.0
wcwidth==0.2.13
